+++ R/arm_ref_comp.R
@@ -3,23 +3,23 @@
 #' @description `r lifecycle::badge("stable")`
 #' Updates the reference and comparison Treatments when the selected Treatment variable changes
 #'
-#' @param session (`environment`) shiny session
-#' @param input (`character`) shiny input
-#' @param output (`character`) shiny input
-#' @param id_ref (`character`) id of reference Treatment input UI element
-#' @param id_comp (`character`) id of comparison group input UI element
-#' @param id_arm_var (`character`) id of Treatment variable input UI element
-#' @param data (`reactive` or `data.frame`) dataset used to validate Treatment reference inputs and
+#' @param session (`environment`)\cr shiny session
+#' @param input (`character`)\cr shiny input
+#' @param output (`character`)\cr shiny input
+#' @param id_ref (`character`)\cr id of reference Treatment input UI element
+#' @param id_comp (`character`)\cr id of comparison group input UI element
+#' @param id_arm_var (`character`)\cr id of Treatment variable input UI element
+#' @param data (`reactive` or `data.frame`)\cr dataset used to validate Treatment reference inputs and
 #'   set `id_ref` input.
-#' @param arm_ref_comp (`unknown`) Treatment reference and compare variables provided as a
+#' @param arm_ref_comp (`unknown`)\cr Treatment reference and compare variables provided as a
 #'   nested list where each Treatment variable corresponds a list specifying the default levels for the
 #'   reference and comparison treatments.
-#' @param module (`character`) name of the module where this is called (this is only used
+#' @param module (`character`)\cr name of the module where this is called (this is only used
 #'   to produce more informative error messages)
-#' @param on_off (`logical`) A reactive that can be used to
-#'   stop the whole observer if FALSE.
-#' @param input_id (`character`) unique id that the buckets will be referenced with.
-#' @param output_id (`character`) name of the UI id that the output will be written to.
+#' @param on_off (`logical`)\cr A reactive that can be used to
+#'   stop the whole observer if `FALSE`.
+#' @param input_id (`character`)\cr unique id that the buckets will be referenced with.
+#' @param output_id (`character`)\cr name of the UI id that the output will be written to.
 #' @return Returns a `shinyvalidate::InputValidator` which checks that there is at least one reference
 #'   and comparison arm
 #' @keywords internal
@@ -139,13 +139,13 @@ arm_ref_comp_observer <- function(session,
 #' Check if the Treatment variable is reference or compare
 #'
 #' @description `r lifecycle::badge("stable")`
-#' @param x (\code{character}) Name of the variable
-#' @param df_to_check (\code{data.frame}) table to check
-#' @param module \code{character} teal module the ref and comp are called in
+#' @param x (`character`)\cr Name of the variable
+#' @param df_to_check (`data.frame`)\cr table to check
+#' @param module (`character`)\cr teal module the ref and comp are called in
 #'
 #' @keywords internal
 #'
-#' @return \code{TRUE} or \code{FALSE} whether the variable is in ref or comp
+#' @return `TRUE` or `FALSE` whether the variable is in ref or comp
 check_arm_ref_comp <- function(x, df_to_check, module) {
   msg <- paste("module", module, "argument arm_ref_comp ")