+++ R/labels.R
@@ -1,8 +1,8 @@
 #' Get full label, useful for annotating plots
 #'
-#' @param dataset \code{data.frame} dataset
-#' @param column \code{character} column to get label from
-#' @param omit_raw_name \code{logical} omits the raw name in square brackets if label is found
+#' @param dataset (`data.frame`)\cr dataset
+#' @param column (`character`)\cr column to get label from
+#' @param omit_raw_name (`logical`)\cr omits the raw name in square brackets if label is found
 #'
 #' @return "Label `[Column name]`" if label exists, otherwise "Column name"
 #'