+++ R/string_ops.R
@@ -3,7 +3,7 @@
 #' Generic to parse text into numeric vectors. This was initially designed
 #' for a robust interpretation of text input in teal modules.
 #'
-#' @param str `vector` to extract numeric from.
+#' @param str (`vector`)\cr to extract numeric from.
 #' @details The function is intended to extract any numeric from a character
 #'   string, factor levels, boolean and return a vector of numeric.
 #'