+++ R/tm_t_pp_basic_info.R
@@ -58,6 +58,7 @@ template_basic_info <- function(dataname = "ANL",
 #' This teal module produces a patient profile basic info report using `ADaM` datasets.
 #'
 #' @inheritParams module_arguments
+#' @inheritParams template_basic_info
 #' @param patient_col (`character`)\cr patient ID column to be used.
 #' @param vars ([teal.transform::choices_selected()] or [teal.transform::data_extract_spec()])\cr
 #' `ADSL` columns to be shown in Basic Info tab.